feat: API 호출에 필요한 프로퍼티와 메서드를 갖는 NetworkManager 구현
- lat, lon, WeatherModel을 갖고, callAPI를 통해 API 호출 데이터 저장
